Accuracy manager
Made a accuracy manager that checks if all inputs for the machine are close to perfect. This will also have a value for the leaderboard/highscore system

+++ Assets/Scripts/Thygo/RPMData.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public struct RPMData
 {
-    public float RPMCurrent; 
+    public int RPMCurrent; 
 }
